refactor: rename modules from PascalCase to kebab-case
