+++ apps/extension/src/contentscript/multitable-panel/assets/styles-dropdown.tsx
