+++ libs/shared-components/src/common/image.tsx
