+++ libs/shared-components/src/mini-overlay/assets/icons/arrow.tsx
