+++ libs/shared-components/src/mini-overlay/assets/icons/bell-with-circle.tsx
