+++ libs/shared-components/src/mini-overlay/assets/icons/bell.tsx
