+++ libs/shared-components/src/mini-overlay/assets/icons/close.tsx
