+++ libs/shared-components/src/mini-overlay/assets/icons/index.ts
@@ -1,9 +1,10 @@
-import MutationFallbackIcon from './MutationFallbackIcon'
-import ArrowIcon from './Arrow'
-import StopTopIcon from './StopTop'
-import PlayCenterIcon from './PlayCenter'
-import StopCenterIcon from './StopCenter'
-import BellIcon from './Bell'
-import BellWithCircle from './BellWithCircle'
+import MutationFallbackIcon from './mutation-fallback-icon'
+import ArrowIcon from './arrow'
+import StopTopIcon from './stop-top'
+import PlayCenterIcon from './play-center'
+import StopCenterIcon from './stop-center'
+import BellIcon from './bell'
+import BellWithCircle from './bell-with-circle'
+import Close from './close'
 
-export { MutationFallbackIcon, ArrowIcon, StopTopIcon, PlayCenterIcon, StopCenterIcon, BellIcon, BellWithCircle }
+export { MutationFallbackIcon, ArrowIcon, StopTopIcon, PlayCenterIcon, StopCenterIcon, BellIcon, BellWithCircle, Close }
