+++ libs/shared-components/src/mini-overlay/assets/icons/mutation-fallback-icon.tsx
