+++ libs/shared-components/src/mini-overlay/assets/icons/play-center.tsx
