+++ libs/shared-components/src/mini-overlay/assets/icons/stop-center.tsx
