+++ libs/shared-components/src/mini-overlay/profile.tsx
