+++ libs/shared-components/src/notifications/assets/icons/blue-badge.tsx
