+++ libs/shared-components/src/notifications/assets/icons/collapse.tsx
