+++ libs/shared-components/src/notifications/assets/icons/green-badge.tsx
