+++ libs/shared-components/src/notifications/assets/icons/notification-close.tsx
