+++ libs/shared-components/src/notifications/assets/icons/notification-message.tsx
