+++ libs/shared-components/src/notifications/assets/icons/red-badge.tsx
